Create Cavafy Impossible Things
Discovered do to Nico Muhly's setting of this to music for the Britten Symphony, performed in Cambridge.  Was amazing.  Used to resonate with me.  Now I think it is hilarious.  Also, there is that tremendous value in the tawdry days full of ennui-- or rather, that if one can learn to be present in them, to appreciate their brevity and beauty.